+++ miden-vm/tests/integration/air/chiplets/bitwise.rs
@@ -6,7 +6,7 @@ fn bitwise_and() {
     let asm_op = "u32and push.0 u32and push.0 u32and push.65535 push.137 u32and";
     let pub_inputs = vec![1, 1];
 
-    build_op_test!(&asm_op, &pub_inputs).prove_and_verify(pub_inputs, false);
+    build_op_test!(&asm_op, &pub_inputs).check_constraints();
 }
 
 #[test]
@@ -15,7 +15,7 @@ fn bitwise_or() {
     let asm_op = "u32or push.0 u32or not push.0 u32or push.65535 push.137 u32or";
     let pub_inputs = vec![1, 1];
 
-    build_op_test!(&asm_op, &pub_inputs).prove_and_verify(pub_inputs, false);
+    build_op_test!(&asm_op, &pub_inputs).check_constraints();
 }
 
 #[test]
@@ -24,7 +24,7 @@ fn bitwise_xor() {
     let asm_op = "u32xor push.0 u32xor push.1 u32xor push.65535 push.137 u32xor";
     let pub_inputs = vec![1, 1];
 
-    build_op_test!(&asm_op, &pub_inputs).prove_and_verify(pub_inputs, false);
+    build_op_test!(&asm_op, &pub_inputs).check_constraints();
 }
 
 #[test]
@@ -32,5 +32,5 @@ fn all_operations() {
     let source = "begin u32and push.0 u32or push.0 u32xor end";
     let pub_inputs = vec![1, 1];
 
-    build_test!(source, &pub_inputs).prove_and_verify(pub_inputs, false);
+    build_test!(source, &pub_inputs).check_constraints();
 }